dev-tools/pyproject.toml
[project]
name = "dev-tools"
version = "1.0"
description = "CLI for making validation logic in PRs"
requires-python = "~=3.13"
dependencies = [
    "marko==2.1.4",
    "pygithub==2.6.1",
    "pytest==9.0.2",
    "pytest-datadir==1.8.0",
    "python-graphql-client==0.4.3",
    "requests==2.32.4",
]

[pytest]
log_cli = true
